OE0-68: integrating mapping with claim mapping
