+++ api_fhir_r4/mapping/__init__.py
@@ -4,4 +4,4 @@ from api_fhir_r4.mapping import activityDefinitionMapping
 from api_fhir_r4.mapping import locationMapping
 from api_fhir_r4.mapping import groupMapping
 from api_fhir_r4.mapping import medicationMapping
-from api_fhir_r4.mapping import claimResponseMapping
+from api_fhir_r4.mapping import claimMapping
